feat(coworker): move posture into the composer lip - header sheds posture (BI-E4CC8E71)
Complete the input-lip convergence deferred in BI-4931A5FE. The coworker posture
(mode / page-editing / web access) leaves the header and lives in the composer
lip next to where you type - matching how Claude Code / Cursor / Cline place
permission controls at the input. The header then sheds posture entirely ->
identity + Skills + overflow only.

Work priority is a separate concern that already lives in its own
CoworkerPriorityDock strip at the composer (#2337), so the new
CoworkerPostureControl holds only mode + page-edit + web - no priority row.

- New CoworkerPostureControl.tsx (+ test): posture summary chip that expands
  into the real switches; opens upward for the bottom-anchored composer; pure UI.
- AgentPanelHeader: sheds the posture block + props (slim header); test trimmed.
- AgentMessageInput: renders the posture control in its lip; composer test
  unchanged (posture not rendered without its handlers).
- AgentCoworkerPanel: moves the posture props header -> composer. Spec section 10.

Deferred earlier as "typing area only" to avoid colliding with the concurrent
golden-triangle composer burst; that has landed, so the convergence is now safe.

UX-Fit-Decision: progressive disclosure (same axis and rationale as BI-706321B1 /
BI-4931A5FE). principle_decide degenerate on human_cognitive_load
(commandmentConflict=false) - decided on merits; the header is simpler, posture
sits where you act, and nothing raw is promoted to the resting surface.
