+++ apps/web/components/agent/AgentPanelHeader.test.tsx
@@ -24,10 +24,6 @@ const baseProps = {
   onConfirmClear: () => {},
   clearDisabled: false,
   clearConfirmOpen: false,
-  elevatedAssistEnabled: false,
-  onToggleElevatedAssist: () => {},
-  externalAccessEnabled: false,
-  onToggleExternalAccess: () => {},
   onClose: () => {},
   onDragStart: () => {},
 };
@@ -35,54 +31,15 @@ const baseProps = {
 afterEach(() => cleanup());
 
 describe("AgentPanelHeader", () => {
-  it("renders a calm resting header: identity, one sensitivity tag, and a Controls summary", () => {
+  it("renders a calm resting header: identity, one sensitivity tag, and an overflow trigger", () => {
     render(<AgentPanelHeader {...baseProps} />);
     expect(screen.getByText("Ops Co-worker")).toBeTruthy();
-    // Session toggles are no longer resting chips — they live behind the posture control.
-    expect(screen.getByText("Controls")).toBeTruthy();
     // The duplicate sensitivity badge is gone: exactly one "Internal" indicator.
     expect(screen.getAllByText("Internal")).toHaveLength(1);
-    // Toggle labels stay hidden until the posture menu is opened.
+    expect(screen.getByRole("button", { name: "More options" })).toBeTruthy();
+    // Posture now lives in the composer lip, not the header.
+    expect(screen.queryByText("Controls")).toBeNull();
     expect(screen.queryByText("Edit fields on this page")).toBeNull();
-  });
-
-  it("summarises the active posture in plain language", () => {
-    render(
-      <AgentPanelHeader
-        {...baseProps}
-        useUnified
-        coworkerMode="act"
-        onToggleCoworkerMode={() => {}}
-        elevatedAssistEnabled
-        externalAccessEnabled
-      />,
-    );
-    expect(screen.getByText("Act · edits on · web on")).toBeTruthy();
-  });
-
-  it("reveals real toggle switches when the posture control is opened", () => {
-    render(<AgentPanelHeader {...baseProps} />);
-    fireEvent.click(screen.getByText("Controls"));
-    const editSwitch = screen.getByRole("switch", { name: "Edit fields on this page" });
-    expect(editSwitch.getAttribute("aria-checked")).toBe("false");
-    expect(screen.getByRole("switch", { name: "Web access" })).toBeTruthy();
-  });
-
-  it("fires the page-editing toggle through the switch", () => {
-    const onToggleElevatedAssist = vi.fn();
-    render(<AgentPanelHeader {...baseProps} onToggleElevatedAssist={onToggleElevatedAssist} />);
-    fireEvent.click(screen.getByText("Controls"));
-    fireEvent.click(screen.getByRole("switch", { name: "Edit fields on this page" }));
-    expect(onToggleElevatedAssist).toHaveBeenCalledTimes(1);
-  });
-
-  it("shows the Advise/Act mode segment only when unified, inside the posture menu", () => {
-    render(
-      <AgentPanelHeader {...baseProps} useUnified coworkerMode="advise" onToggleCoworkerMode={() => {}} />,
-    );
-    // The resting summary reads "Advise"; clicking it opens the posture menu.
-    fireEvent.click(screen.getByText("Advise"));
-    expect(screen.getAllByText("Act").length).toBeGreaterThan(0);
   });
 
   it("moves profile, diagnostics, and erase into the overflow menu and points dev work to Build Studio", () => {
@@ -97,7 +54,6 @@ describe("AgentPanelHeader", () => {
         onOpenClearConfirm={onOpenClearConfirm}
       />,
     );
-    // Secondary actions are not in the resting header.
     expect(screen.queryByText("Erase conversation")).toBeNull();
     fireEvent.click(screen.getByRole("button", { name: "More options" }));
     expect(screen.getByText("View profile, skills & tools")).toBeTruthy();
