+++ apps/web/components/agent/AgentPanelHeader.tsx
@@ -24,10 +24,6 @@ type Props = {
   onConfirmClear: () => void;
   clearDisabled: boolean;
   clearConfirmOpen: boolean;
-  elevatedAssistEnabled: boolean;
-  onToggleElevatedAssist: () => void;
-  externalAccessEnabled: boolean;
-  onToggleExternalAccess: () => void;
   onClose: () => void;
   onDragStart: (e: React.MouseEvent) => void;
   onViewProfile?: () => void;
@@ -35,86 +31,9 @@ type Props = {
   devMode?: boolean;
   canUseDev?: boolean;
   onToggleDev?: () => void;
-  coworkerMode?: "advise" | "act";
-  onToggleCoworkerMode?: () => void;
-  useUnified?: boolean;
   marketingSkillRules?: Record<string, { visible?: boolean; label?: string; reframe?: string }> | null;
   isDocked?: boolean;
 };
-
-/** A labelled row with a real switch affordance, used inside the posture menu. */
-function ToggleSwitchRow({
-  checked,
-  onToggle,
-  label,
-  description,
-  title,
-}: {
-  checked: boolean;
-  onToggle: () => void;
-  label: string;
-  description: string;
-  title: string;
-}) {
-  return (
-    <button
-      type="button"
-      role="switch"
-      aria-checked={checked}
-      aria-label={label}
-      title={title}
-      onMouseDown={(e) => e.stopPropagation()}
-      onClick={(e) => {
-        e.stopPropagation();
-        onToggle();
-      }}
-      style={{
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "space-between",
-        gap: 12,
-        width: "100%",
-        textAlign: "left",
-        background: "none",
-        border: "none",
-        padding: "8px 12px",
-        cursor: "pointer",
-      }}
-    >
-      <span style={{ display: "flex", flexDirection: "column", gap: 1, minWidth: 0 }}>
-        <span style={{ fontSize: 12, color: "var(--dpf-text)" }}>{label}</span>
-        <span style={{ fontSize: 10, color: "var(--dpf-muted)", lineHeight: 1.3 }}>{description}</span>
-      </span>
-      <span
-        aria-hidden="true"
-        style={{
-          position: "relative",
-          width: 34,
-          height: 18,
-          borderRadius: 999,
-          flex: "0 0 auto",
-          background: checked
-            ? "var(--dpf-success)"
-            : "color-mix(in srgb, var(--dpf-text) 22%, transparent)",
-          transition: "background 0.15s ease",
-        }}
-      >
-        <span
-          style={{
-            position: "absolute",
-            top: 2,
-            left: checked ? 18 : 2,
-            width: 14,
-            height: 14,
-            borderRadius: "50%",
-            background: "var(--dpf-text)",
-            transition: "left 0.15s ease",
-          }}
-        />
-      </span>
-    </button>
-  );
-}
 
 /** A single action row inside the overflow menu. */
 function MenuActionRow({
@@ -175,59 +94,28 @@ export function AgentPanelHeader({
   onConfirmClear,
   clearDisabled,
   clearConfirmOpen,
-  elevatedAssistEnabled,
-  onToggleElevatedAssist,
-  externalAccessEnabled,
-  onToggleExternalAccess,
   onClose,
   onDragStart,
   providerInfo,
   devMode,
   canUseDev,
   onToggleDev,
-  coworkerMode,
-  onToggleCoworkerMode,
   onViewProfile,
-  useUnified,
   marketingSkillRules,
   isDocked = false,
 }: Props) {
   const rootRef = useRef<HTMLDivElement>(null);
-  const [menu, setMenu] = useState<null | "posture" | "more">(null);
-  const showMode = Boolean(useUnified && onToggleCoworkerMode);
-  const isAct = coworkerMode === "act";
+  const [menuOpen, setMenuOpen] = useState(false);
 
-  // Close any open menu on a click outside the header.
+  // Close the overflow menu on a click outside the header.
   useEffect(() => {
-    if (!menu) return;
+    if (!menuOpen) return;
     function onDocMouseDown(e: MouseEvent) {
-      if (rootRef.current && !rootRef.current.contains(e.target as Node)) setMenu(null);
+      if (rootRef.current && !rootRef.current.contains(e.target as Node)) setMenuOpen(false);
     }
     document.addEventListener("mousedown", onDocMouseDown);
     return () => document.removeEventListener("mousedown", onDocMouseDown);
-  }, [menu]);
-
-  // Plain-language summary of the active posture for the resting chip.
-  const summaryParts: string[] = [];
-  if (showMode) summaryParts.push(isAct ? "Act" : "Advise");
-  if (elevatedAssistEnabled) summaryParts.push("edits on");
-  if (externalAccessEnabled) summaryParts.push("web on");
-  const summaryLabel = summaryParts.length > 0 ? summaryParts.join(" · ") : "Controls";
-  const postureActive = elevatedAssistEnabled || externalAccessEnabled || (showMode && isAct);
-
-  const popoverShell: React.CSSProperties = {
-    position: "absolute",
-    top: "calc(100% + 8px)",
-    right: 0,
-    width: 248,
-    maxWidth: "calc(100vw - 32px)",
-    background: "var(--dpf-surface-1)",
-    border: "1px solid var(--dpf-border)",
-    borderRadius: 12,
-    boxShadow: "0 10px 28px rgba(0,0,0,0.35)",
-    zIndex: 5,
-    padding: "6px 0",
-  };
+  }, [menuOpen]);
 
   return (
     <div
@@ -308,147 +196,15 @@ export function AgentPanelHeader({
 
       {/* Right control cluster */}
       <div style={{ display: "flex", alignItems: "center", gap: 6, position: "relative", flex: "0 0 auto" }}>
-        {/* Posture control */}
-        <div style={{ position: "relative" }} onMouseDown={(e) => e.stopPropagation()}>
-          <button
-            type="button"
-            onClick={(e) => {
-              e.stopPropagation();
-              setMenu((m) => (m === "posture" ? null : "posture"));
-            }}
-            aria-expanded={menu === "posture"}
-            aria-haspopup="dialog"
-            title="Conversation controls — mode, page editing, web access, and priority for this coworker"
-            style={{
-              display: "inline-flex",
-              alignItems: "center",
-              gap: 5,
-              fontSize: 10,
-              color: postureActive ? "var(--dpf-text)" : "var(--dpf-muted)",
-              background: postureActive ? "color-mix(in srgb, var(--dpf-accent) 12%, transparent)" : "transparent",
-              border: `1px solid ${postureActive ? "color-mix(in srgb, var(--dpf-accent) 45%, transparent)" : "var(--dpf-border)"}`,
-              borderRadius: 999,
-              padding: "3px 8px",
-              cursor: "pointer",
-              lineHeight: 1.2,
-              maxWidth: 180,
-            }}
-          >
-            {postureActive && (
-              <span
-                aria-hidden="true"
-                style={{ width: 6, height: 6, borderRadius: "50%", background: "var(--dpf-accent)", flex: "0 0 auto" }}
-              />
-            )}
-            <span style={{ overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }}>{summaryLabel}</span>
-            <span aria-hidden="true" style={{ fontSize: 8, opacity: 0.8 }}>▾</span>
-          </button>
-          {menu === "posture" && (
-            <div role="dialog" aria-label="Conversation controls" style={popoverShell}>
-              <div
-                style={{
-                  fontSize: 9,
-                  textTransform: "uppercase",
-                  letterSpacing: "0.06em",
-                  color: "var(--dpf-muted)",
-                  padding: "4px 12px 6px",
-                }}
-              >
-                This conversation
-              </div>
-              {showMode && (
-                <div
-                  style={{
-                    display: "flex",
-                    alignItems: "center",
-                    justifyContent: "space-between",
-                    gap: 12,
-                    padding: "6px 12px",
-                  }}
-                >
-                  <span style={{ fontSize: 12, color: "var(--dpf-text)" }}>Mode</span>
-                  <span
-                    style={{
-                      display: "inline-flex",
-                      border: "1px solid var(--dpf-border)",
-                      borderRadius: 999,
-                      overflow: "hidden",
-                    }}
-                  >
-                    {(["advise", "act"] as const).map((m) => {
-                      const active = isAct ? m === "act" : m === "advise";
-                      return (
-                        <button
-                          key={m}
-                          type="button"
-                          onMouseDown={(e) => e.stopPropagation()}
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            if (!active) onToggleCoworkerMode?.();
-                          }}
-                          title={
-                            m === "act"
-                              ? "Act: the coworker executes within your authority"
-                              : "Advise: the coworker recommends but doesn't act"
-                          }
-                          style={{
-                            fontSize: 11,
-                            padding: "2px 10px",
-                            border: "none",
-                            cursor: active ? "default" : "pointer",
-                            fontWeight: active ? 600 : 400,
-                            color: active ? "var(--dpf-text)" : "var(--dpf-muted)",
-                            background: active
-                              ? m === "act"
-                                ? "color-mix(in srgb, var(--dpf-success) 18%, transparent)"
-                                : "var(--dpf-surface-2)"
-                              : "transparent",
-                          }}
-                        >
-                          {m === "act" ? "Act" : "Advise"}
-                        </button>
-                      );
-                    })}
-                  </span>
-                </div>
-              )}
-              <ToggleSwitchRow
-                checked={elevatedAssistEnabled}
-                onToggle={onToggleElevatedAssist}
-                label="Edit fields on this page"
-                description="Let the coworker fill in forms for you"
-                title={
-                  elevatedAssistEnabled
-                    ? "On: this page's coworker can update approved form fields"
-                    : "Off: this page's coworker only suggests changes"
-                }
-              />
-              <ToggleSwitchRow
-                checked={externalAccessEnabled}
-                onToggle={onToggleExternalAccess}
-                label="Web access"
-                description="Allow web search and fetch this session"
-                title={
-                  externalAccessEnabled
-                    ? "On: this page's coworker can use approved public web search and fetch tools"
-                    : "Off: this page's coworker cannot reach public web tools"
-                }
-              />
-              {/* Priority moved out of this menu — it now lives in-flow at the composer
-                  (CoworkerPriorityDock), so the triangle stays in view and nothing clips. */}
-            </div>
-          )}
-        </div>
-
         {/* Overflow menu */}
         <div style={{ position: "relative" }} onMouseDown={(e) => e.stopPropagation()}>
           <button
             type="button"
             onClick={(e) => {
               e.stopPropagation();
-              setMenu((m) => (m === "more" ? null : "more"));
+              setMenuOpen((m) => !m);
             }}
-            aria-expanded={menu === "more"}
+            aria-expanded={menuOpen}
             aria-haspopup="menu"
             aria-label="More options"
             title="More — profile, diagnostics, erase conversation"
@@ -469,14 +225,30 @@ export function AgentPanelHeader({
           >
             ⋯
           </button>
-          {menu === "more" && (
-            <div role="menu" aria-label="More options" style={popoverShell}>
+          {menuOpen && (
+            <div
+              role="menu"
+              aria-label="More options"
+              style={{
+                position: "absolute",
+                top: "calc(100% + 8px)",
+                right: 0,
+                width: 248,
+                maxWidth: "calc(100vw - 32px)",
+                background: "var(--dpf-surface-1)",
+                border: "1px solid var(--dpf-border)",
+                borderRadius: 12,
+                boxShadow: "0 10px 28px rgba(0,0,0,0.35)",
+                zIndex: 5,
+                padding: "6px 0",
+              }}
+            >
               {onViewProfile && (
                 <MenuActionRow
                   label="View profile, skills & tools"
                   title="View coworker profile, skills, and tools"
                   onClick={() => {
-                    setMenu(null);
+                    setMenuOpen(false);
                     onViewProfile();
                   }}
                 />
@@ -491,7 +263,7 @@ export function AgentPanelHeader({
                       : "Diagnostics: inspect page context and explain likely fixes. Use Build Studio for code-changing work"
                   }
                   onClick={() => {
-                    setMenu(null);
+                    setMenuOpen(false);
                     onToggleDev();
                   }}
                 />
@@ -502,7 +274,7 @@ export function AgentPanelHeader({
                 disabled={clearDisabled}
                 title="Erase current conversation"
                 onClick={() => {
-                  setMenu(null);
+                  setMenuOpen(false);
                   onOpenClearConfirm();
                 }}
               />
